Moved INIManager to engine.ini subpackage
